The newInstance method was added to NewsFragment and ProfileFragment, the Toast was deleted and the colors and dimensions were extracted to resources
