+++ app/src/main/java/ar/com/wolox/android/example/ui/home/HomeActivity.kt
@@ -1,7 +1,6 @@
 package ar.com.wolox.android.example.ui.home
 
 import ar.com.wolox.android.R
-import ar.com.wolox.android.example.ui.home.fragment.HomeFragment
 import ar.com.wolox.wolmo.core.activity.WolmoActivity
 import dagger.Lazy
 import javax.inject.Inject
