+++ app/src/main/java/ar/com/wolox/android/example/ui/home/HomeModule.kt
@@ -1,6 +1,5 @@
 package ar.com.wolox.android.example.ui.home
 
-import ar.com.wolox.android.example.ui.home.fragment.HomeFragment
 import ar.com.wolox.android.example.ui.home.news.NewsFragment
 import ar.com.wolox.android.example.ui.home.profile.ProfileFragment
 import dagger.Module
