+++ app/src/main/java/ar/com/wolox/android/example/ui/home/profile/ProfilePresenter.kt
@@ -8,5 +8,4 @@ class ProfilePresenter @Inject constructor(private val mRetrofitServices: Retrof
 
     override fun onViewAttached() {
     }
-
 }
